Rest Documentation Implementation

+++ src/app/pages/rest-documentation/rest-documentation.component.ts
@@ -19,12 +19,14 @@ export class RestDocumentationComponent extends PageComponent implements OnInit 
 
   dto: RestDocumentationDto = new RestDocumentationDto();
   serverUrl = '';
+  restDocumentationEndpoint: RestDocumentationEndpoint;
+  isDeleteCompleted = false;
   constructor(private service: RestDocumentationService,
     private activatedRoute: ActivatedRoute,
     private httpClient: HttpClient,
             private navigatorService: CommandNavigatorService,
               private dynamicRequestService:  DynamicRequestService,
-              private sanitizer: DomSanitizer) {
+              private sanitizer: DomSanitizer,) {
   super();
 }
 
@@ -39,6 +41,15 @@ export class RestDocumentationComponent extends PageComponent implements OnInit 
     const locateParams = this.getLocateParams();
     this.service.getById(locateParams.get('ID')).subscribe(data => {
       this.dto = data;
+
+      this.dto.restDocumentationEndpoints
+      .filter(endpoint => endpoint.type === "form")
+      .filter(endpoint => ["post", "put"].includes(endpoint.method))
+      .forEach(endpoint => {
+        endpoint.jsonString = atob(endpoint.jsonString); 
+        const jsonObj = JSON.parse(endpoint.jsonString);
+        endpoint.jsonString = JSON.stringify(jsonObj, null, 4);
+      })
     });
   }
 
@@ -58,12 +69,23 @@ export class RestDocumentationComponent extends PageComponent implements OnInit 
     return this.sanitizer.bypassSecurityTrustHtml(resource);
   }
 
-  runDelete(restDocumentationEndpoint: RestDocumentationEndpoint) {
+  runDelete(restDocumentationEndpoint) {
+
+    if (!restDocumentationEndpoint.selectionId) {
+      alert("Selection ID parameter is missing.");
+      return;
+    }
+
     const url = '/dataset/' + restDocumentationEndpoint.form.jsonUrl + '?selection-id=' + restDocumentationEndpoint.selectionId;
-  
+
     this.dynamicRequestService.deleteFromBackend(url).subscribe(
       (data) => {
         restDocumentationEndpoint.restResults = JSON.stringify(data, null, 4);
+        this.isDeleteCompleted = true;
+
+        setTimeout(() => {
+          this.isDeleteCompleted = false;
+        }, 3000); 
       },
       (error) => {
         console.error('Error deleting data:', error);
@@ -71,14 +93,50 @@ export class RestDocumentationComponent extends PageComponent implements OnInit 
     );
   }
 
-  runPost(){
-
+  runPost(restDocumentationEndpoint: RestDocumentationEndpoint) {
+    const url = '/dataset/'+restDocumentationEndpoint.form.jsonUrl + '?';;
+  
+  
+    const requestBody = restDocumentationEndpoint.jsonString;
+      
+    
+    this.dynamicRequestService.postToBackendWithTextResponse(url, requestBody).subscribe(
+      (id) => {
+        alert(id);
+        restDocumentationEndpoint.restResults = id;
+      },
+      (error) => {
+        console.error('Error making POST request:', error);
+      }
+    );
   }
 
 
   runPut(){
 
 
+  }
+
+ 
+  getJsonUrlValue(restDocumentationEndpoint: RestDocumentationEndpoint): string {
+    if (restDocumentationEndpoint.list) {
+      return this.getListJsonUrl(restDocumentationEndpoint);
+    } else {
+      return this.getFormJsonUrl(restDocumentationEndpoint);
+    }
+  }
+  
+  getListJsonUrl(restDocumentationEndpoint: RestDocumentationEndpoint): string {
+    return this.serverUrl + '/datalist/' + restDocumentationEndpoint.list?.jsonUrl;
+  }
+  
+  getFormJsonUrl(restDocumentationEndpoint: RestDocumentationEndpoint): string {
+    const method = restDocumentationEndpoint.method;
+    if (method === 'delete') {
+      return this.serverUrl + '/dataset/' + restDocumentationEndpoint.form?.jsonUrl + '?selection-id={id}';
+    } else {
+      return this.serverUrl + '/dataset/' + restDocumentationEndpoint.form?.jsonUrl;
+    }
   }
 
 }
